requirements.txt
aiofiles==24.1.0
asttokens==2.4.1
attrs==24.2.0
bech32==1.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cosmpy==0.9.2
ecdsa==0.19.0
Flask==3.1.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
nada_dsl==0.7.1
nillion-python-helpers==0.3.1
packaging==24.2
parsial==0.1.0
pluggy==1.5.0
priority==2.0.0
protobuf==4.25.5
py-nillion-client==0.6.0
pycryptodome==3.21.0
pytest==8.3.3
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
Quart==0.19.9
referencing==0.35.1
requests==2.32.3
richreports==0.2.0
rpds-py==0.21.0
six==1.16.0
sortedcontainers==2.4.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
